adapt new strategy for uploading
