+++ modules/renter/files.go
@@ -19,6 +19,7 @@ type File struct {
 // of the file piece.
 type FilePiece struct {
 	Active     bool                     // Set to true if the host is online and has the file, false otherwise.
+	Repairing  bool                     // Set to true if there's an upload happening for the piece at the moment.
 	Contract   consensus.FileContract   // The contract being enforced.
 	ContractID consensus.FileContractID // The ID of the contract.
 	HostIP     modules.NetAddress       // Where to find the file.
@@ -46,7 +47,15 @@ func (f *File) Nickname() string {
 
 // Repairing returns whether or not the file is actively being repaired.
 func (f *File) Repairing() bool {
-	return false // Repairing is not currently supported.
+	f.renter.mu.RLock()
+	defer f.renter.mu.RUnlock()
+
+	for _, piece := range f.pieces {
+		if piece.Repairing {
+			return true
+		}
+	}
+	return false
 }
 
 // TimeRemaining returns the amount of time until the file's contracts expire.
